+++ WebApi/Models/WebShippingDetail.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Security.Permissions;
 using System.Text;
 using System.Threading.Tasks;
@@ -9,8 +10,8 @@ namespace WebApi.Models
 {
     public class WebShippingDetail
     {
-        public int Id { get; set; }
-        public int VehicleId { get; set; }
+        public BigInteger Id { get; set; }
+        public BigInteger VehicleId { get; set; }
         public string VehicleName { get; set; }
         public double VehiclePrice { get; set; }
         public string VehicleDescription { get; set;}
